Sql support - wip
