+++ xtask/Cargo.toml
@@ -11,3 +11,4 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sysinfo = "0.28"
 anyhow = "1.0"
+futures = "0.3"
