--- xtask/src/task/mod.rs
@@ -1,30 +1,0 @@
-use anyhow::Result;
-use std::time::Duration;
-
-pub(crate) mod compile;
-pub(crate) mod run;
-
-pub(crate) trait Task {
-    /// Returns the name of the task being benchmarked
-    fn get_name(&self) -> String;
-    /// Executes any actions required before time measurement starts
-    /// By default we make sure we have a release build ready
-    fn prepare(&mut self) -> Result<()> {
-        Ok(())
-    }
-    /// Executes the task to be measured and returns the time it took
-    fn execute(&self) -> Result<Duration>;
-
-    /// Benchmarks the current task and returns the avarage execution time
-    fn benchmark(&mut self, executions: u32) -> Result<Duration> {
-        self.prepare()?;
-        //Cold run
-        self.execute()?;
-        let mut duration = Duration::from_millis(0);
-        for _ in 0..executions {
-            duration += self.execute()?;
-        }
-
-        Ok(duration / executions)
-    }
-}
